+++ src/actions/actions.js
@@ -1,3 +1,0 @@
-import * as authActionCreator from './auth';
-
-export { authActionCreator };