+++ src/actions/auth.js
@@ -1,8 +1,0 @@
-import { authTypes } from '../constants/actionTypes';
-
-export function loginRequest(payload) {
-	return {
-		type: authTypes.LOGIN_REQUEST,
-		payload
-	};
-}