+++ src/api/authApi.js
@@ -1,10 +1,8 @@
-import axios from 'axios';
-
-import { authInstance } from '../axios/axiosGlobals';
+import { axiosInstance } from "../axios";
 
 export function userLogin(params) {
-	return authInstance
-		.post(`/login`, params)
-		.then((response) => ({ params, response }))
-		.catch((error) => ({ params, error }));
+  return axiosInstance
+    .post(`/login`, params)
+    .then(response => ({ params, response }))
+    .catch(error => ({ params, error }));
 }