+++ src/axios/index.js
@@ -1,0 +1,3 @@
+import { axiosInstance } from "./instances";
+
+export { axiosInstance };