+++ src/constants/apiCredentials.js
@@ -1,4 +1,0 @@
-export const BASE_URL = '';
-export const APP_ID = '';
-export const API_KEY = '';
-export const JWT_KEY = '';