+++ src/polyfills.js
@@ -1,2 +1,0 @@
-import 'core-js/fn/array/find';
-import 'core-js/fn/array/includes';