+++ src/routes/AppRoutes.js
@@ -1,19 +1,15 @@
-import React from 'react';
-import { Router, Route, Switch } from 'react-router-dom';
-import history from './history';
+import React from "react";
+import { Router, Route, Switch } from "react-router-dom";
+import history from "./history";
 
-import App from '../App';
-import LoginContainer from '../components/login/LoginContainer';
-import PageNotFound from '../commons/404Page';
+import App from "../App";
 
-const AppRoutes = ({ store }) => (
-	<Router history={history}>
-		<Switch>
-			<Route exact path="/" component={App} />
-			<Route path="/login" component={LoginContainer} />
-			<Route component={PageNotFound} />
-		</Switch>
-	</Router>
+const AppRoutes = () => (
+  <Router history={history}>
+    <Switch>
+      <Route exact path="/" component={App} />
+    </Switch>
+  </Router>
 );
 
 export default AppRoutes;