+++ src/routes/index.js
@@ -1,0 +1,4 @@
+import AppRoutes from "./appRoutes";
+import history from "./history";
+
+export { AppRoutes, history };