+++ src/sagas/authSaga.js
@@ -1,19 +1,11 @@
-import { call, takeLatest, put } from 'redux-saga/effects';
+import { call, takeLatest, put } from "redux-saga/effects";
 
-import { authApi } from '../api/index';
-import { makeAPICall } from './rootSaga';
+import { authTypes } from "../constants/actionTypes";
+import { authApi } from "../api/index";
+import { rootWorkerSaga, actionTypeFormatter } from "./rootSaga";
 
 const authSaga = [
-	takeLatest('LOGIN_REQUEST', prepareAPICall, authApi.userLogin)
+  takeLatest(authTypes.user.login.request, rootWorkerSaga, authApi.userLogin)
 ];
-
-function prepareAPICall(apifunc, action) {
-	makeAPICall(
-		action,
-		apifunc,
-		`${action.type}_SUCCESS`,
-		`${action.type}_FAILURE`
-	);
-}
 
 export default authSaga;